some minor fixes
*   re-organize cmake build rules;
*   fix bugs in log_apple.mm;
