--- Firestore/core/src/firebase/firestore/public/log.h
@@ -1,42 +1,0 @@
-/*
- * Copyright 2017 Google
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-#ifndef FIRESTORE_CORE_SRC_FIREBASE_FIRESTORE_PUBLIC_LOG_H_
-#define FIRESTORE_CORE_SRC_FIREBASE_FIRESTORE_PUBLIC_LOG_H_
-
-namespace firebase {
-namespace firestore {
-
-/// @brief Levels used when logging messages.
-enum LogLevel {
-  /// Verbose Log Level
-  kLogLevelVerbose = 0,
-  /// Debug Log Level
-  kLogLevelDebug,
-  /// Info Log Level
-  kLogLevelInfo,
-  /// Warning Log Level
-  kLogLevelWarning,
-  /// Error Log Level
-  kLogLevelError,
-  /// Assert Log Level
-  kLogLevelAssert,
-};
-
-}  // namespace firestore
-}  // namespace firebase
-
-#endif  // FIRESTORE_CORE_SRC_FIREBASE_FIRESTORE_PUBLIC_LOG_H_
